#690 Upcoming cards
Refactor to common enums and utils
