+++ app/src/main/java/it/niedermann/nextcloud/deck/ui/upcomingcards/UpcomingCardsAdapter.java
@@ -1,21 +1,15 @@
 package it.niedermann.nextcloud.deck.ui.upcomingcards;
 
 import android.app.Activity;
-import android.content.Context;
 import android.view.LayoutInflater;
 import android.view.ViewGroup;
 
 import androidx.annotation.ColorInt;
 import androidx.annotation.NonNull;
-import androidx.annotation.Nullable;
-import androidx.annotation.StringRes;
 import androidx.core.content.ContextCompat;
 import androidx.fragment.app.FragmentManager;
 import androidx.recyclerview.widget.RecyclerView;
 
-import java.time.Instant;
-import java.time.LocalDate;
-import java.time.ZoneId;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.Comparator;
@@ -37,7 +31,7 @@ import it.niedermann.nextcloud.deck.ui.card.DefaultCardViewHolder;
 import it.niedermann.nextcloud.deck.ui.card.EditActivity;
 
 import static androidx.preference.PreferenceManager.getDefaultSharedPreferences;
-import static java.time.temporal.ChronoUnit.DAYS;
+import static it.niedermann.nextcloud.deck.ui.upcomingcards.UpcomingCardsUtil.getDueType;
 
 public class UpcomingCardsAdapter extends RecyclerView.Adapter<RecyclerView.ViewHolder> {
 
@@ -156,12 +150,13 @@ public class UpcomingCardsAdapter extends RecyclerView.Adapter<RecyclerView.View
             }
             return null;
         }), Comparator.nullsLast(Comparator.naturalOrder()));
+
         comparator.thenComparing(card -> {
             if (card != null &&
                     card.getFullCard() != null &&
                     card.getFullCard().getCard().getDueDate() != null) {
 
-                Card c = card.getFullCard().getCard();
+                final Card c = card.getFullCard().getCard();
 
                 if (c.getLastModified() == null && c.getLastModifiedLocal() != null) {
                     return c.getLastModifiedLocal();
@@ -179,6 +174,7 @@ public class UpcomingCardsAdapter extends RecyclerView.Adapter<RecyclerView.View
                 items,
                 comparator
         );
+
         EUpcomingDueType lastDueType = null;
         for (UpcomingCardsAdapterItem filterWidgetCard : items) {
             final EUpcomingDueType nextDueType = getDueType(filterWidgetCard.getFullCard().getCard().getDueDate());
@@ -191,52 +187,5 @@ public class UpcomingCardsAdapter extends RecyclerView.Adapter<RecyclerView.View
         }
 
         notifyDataSetChanged();
-    }
-
-    @NonNull
-    private static EUpcomingDueType getDueType(@Nullable Instant dueDate) {
-        if (dueDate == null) {
-            return EUpcomingDueType.NO_DUE;
-        }
-
-        long diff = DAYS.between(LocalDate.now(), dueDate.atZone(ZoneId.systemDefault()).toLocalDate());
-
-        if (diff > 7) {
-            return EUpcomingDueType.LATER;
-        } else if (diff > 1) {
-            return EUpcomingDueType.WEEK;
-        } else if (diff > 0) {
-            return EUpcomingDueType.TOMORROW;
-        } else if (diff == 0) {
-            return EUpcomingDueType.TODAY;
-        } else {
-            return EUpcomingDueType.OVERDUE;
-        }
-    }
-
-    private enum EUpcomingDueType {
-        OVERDUE(1, R.string.filter_overdue),
-        TODAY(2, R.string.filter_today),
-        TOMORROW(3, R.string.filter_tomorrow),
-        WEEK(4, R.string.filter_week),
-        LATER(5, R.string.filter_later),
-        NO_DUE(6, R.string.filter_no_due);
-
-        private final int value;
-        private final int id;
-
-        EUpcomingDueType(int id, @StringRes int value) {
-            this.value = value;
-            this.id = id;
-        }
-
-        public int getId() {
-            return id;
-        }
-
-        @NonNull
-        public String toString(Context context) {
-            return context.getString(this.value);
-        }
     }
 }
